fix find_templates, update generate_menu logic
when we had files in different directories with same names,
`find_templates` would use only the first path for all of them. This
commit fixes that and it will have to be manually copied to master.

`generate_menu` function now displays the tempalte's path instead of the
path we're creating the file to